Quality of Life refactor
Changes a bunch of internal and external APIs to clean up the IModMetadata/IModVersion split. IModMetadata should be used when talking about a mod in the general sense, and IModVersion should be used when discussing a specific version of the mod.

Also cleans up some more domain handling code for preparation of custom domains later on. Planned to break this apart more later.

This commit also adds basic support for checksums after a mod has been downloaded from Curseforge. It isn't used here yet, but it can be used later to check if the mod has already been downloaded, and may be skipped later.