Use checksum to skip existing files
The checksum that is acquired when a mod finishes downloading is now used to check if the mod has already been downloaded, and not tampered with.

May be used in the future to verify that a mod has downloaded successfully by asking the user to verify it against a known hash.